testutil: move contains/deepcontains checker to separate file
