+++ testutil/checkers_test.go
@@ -32,7 +32,6 @@ import (
 	"path/filepath"
 	"reflect"
 	"regexp"
-	"runtime"
 	"testing"
 
 	"gopkg.in/check.v1"
@@ -68,197 +67,6 @@ func testCheck(c *check.C, checker check.Checker, result bool, error string, par
 			info.Name, params, resultActual, errorActual, result, error)
 	}
 	return params, names
-}
-
-func (s *CheckersS) TestUnsupportedTypes(c *check.C) {
-	testInfo(c, Contains, "Contains", []string{"container", "elem"})
-	testCheck(c, Contains, false, "int is not a supported container", 5, nil)
-	testCheck(c, Contains, false, "bool is not a supported container", false, nil)
-	testCheck(c, Contains, false, "element is a int but expected a string", "container", 1)
-}
-
-func (s *CheckersS) TestContainsVerifiesTypes(c *check.C) {
-	testInfo(c, Contains, "Contains", []string{"container", "elem"})
-	testCheck(c, Contains,
-		false, "container has items of type int but expected element is a string",
-		[...]int{1, 2, 3}, "foo")
-	testCheck(c, Contains,
-		false, "container has items of type int but expected element is a string",
-		[]int{1, 2, 3}, "foo")
-	// This looks tricky, Contains looks at _values_, not at keys
-	testCheck(c, Contains,
-		false, "container has items of type int but expected element is a string",
-		map[string]int{"foo": 1, "bar": 2}, "foo")
-	testCheck(c, Contains,
-		false, "container has items of type int but expected element is a string",
-		map[string]int{"foo": 1, "bar": 2}, "foo")
-}
-
-type animal interface {
-	Sound() string
-}
-
-type dog struct{}
-
-func (d *dog) Sound() string {
-	return "bark"
-}
-
-type cat struct{}
-
-func (c *cat) Sound() string {
-	return "meow"
-}
-
-type tree struct{}
-
-func (s *CheckersS) TestContainsVerifiesInterfaceTypes(c *check.C) {
-	testCheck(c, Contains,
-		false, "container has items of interface type testutil.animal but expected element does not implement it",
-		[...]animal{&dog{}, &cat{}}, &tree{})
-	testCheck(c, Contains,
-		false, "container has items of interface type testutil.animal but expected element does not implement it",
-		[]animal{&dog{}, &cat{}}, &tree{})
-	testCheck(c, Contains,
-		false, "container has items of interface type testutil.animal but expected element does not implement it",
-		map[string]animal{"dog": &dog{}, "cat": &cat{}}, &tree{})
-}
-
-func (s *CheckersS) TestContainsString(c *check.C) {
-	c.Assert("foo", Contains, "f")
-	c.Assert("foo", Contains, "fo")
-	c.Assert("foo", check.Not(Contains), "foobar")
-}
-
-type myString string
-
-func (s *CheckersS) TestContainsCustomString(c *check.C) {
-	c.Assert(myString("foo"), Contains, myString("f"))
-	c.Assert(myString("foo"), Contains, myString("fo"))
-	c.Assert(myString("foo"), check.Not(Contains), myString("foobar"))
-	c.Assert("foo", Contains, myString("f"))
-	c.Assert("foo", Contains, myString("fo"))
-	c.Assert("foo", check.Not(Contains), myString("foobar"))
-	c.Assert(myString("foo"), Contains, "f")
-	c.Assert(myString("foo"), Contains, "fo")
-	c.Assert(myString("foo"), check.Not(Contains), "foobar")
-}
-
-func (s *CheckersS) TestContainsArray(c *check.C) {
-	c.Assert([...]int{1, 2, 3}, Contains, 1)
-	c.Assert([...]int{1, 2, 3}, Contains, 2)
-	c.Assert([...]int{1, 2, 3}, Contains, 3)
-	c.Assert([...]int{1, 2, 3}, check.Not(Contains), 4)
-	c.Assert([...]animal{&dog{}, &cat{}}, Contains, &dog{})
-	c.Assert([...]animal{&cat{}}, check.Not(Contains), &dog{})
-}
-
-func (s *CheckersS) TestContainsSlice(c *check.C) {
-	c.Assert([]int{1, 2, 3}, Contains, 1)
-	c.Assert([]int{1, 2, 3}, Contains, 2)
-	c.Assert([]int{1, 2, 3}, Contains, 3)
-	c.Assert([]int{1, 2, 3}, check.Not(Contains), 4)
-	c.Assert([]animal{&dog{}, &cat{}}, Contains, &dog{})
-	c.Assert([]animal{&cat{}}, check.Not(Contains), &dog{})
-}
-
-func (s *CheckersS) TestContainsMap(c *check.C) {
-	c.Assert(map[string]int{"foo": 1, "bar": 2}, Contains, 1)
-	c.Assert(map[string]int{"foo": 1, "bar": 2}, Contains, 2)
-	c.Assert(map[string]int{"foo": 1, "bar": 2}, check.Not(Contains), 3)
-	c.Assert(map[string]animal{"dog": &dog{}, "cat": &cat{}}, Contains, &dog{})
-	c.Assert(map[string]animal{"cat": &cat{}}, check.Not(Contains), &dog{})
-}
-
-// Arbitrary type that is not comparable
-type myStruct struct {
-	attrs map[string]string
-}
-
-func (s *CheckersS) TestContainsUncomparableType(c *check.C) {
-	if runtime.Compiler != "go" {
-		c.Skip("this test only works on go (not gccgo)")
-	}
-
-	elem := myStruct{map[string]string{"k": "v"}}
-	containerArray := [...]myStruct{elem}
-	containerSlice := []myStruct{elem}
-	containerMap := map[string]myStruct{"foo": elem}
-	errMsg := "runtime error: comparing uncomparable type testutil.myStruct"
-	testInfo(c, Contains, "Contains", []string{"container", "elem"})
-	testCheck(c, Contains, false, errMsg, containerArray, elem)
-	testCheck(c, Contains, false, errMsg, containerSlice, elem)
-	testCheck(c, Contains, false, errMsg, containerMap, elem)
-}
-
-func (s *CheckersS) TestDeepContainsUnsupportedTypes(c *check.C) {
-	testInfo(c, DeepContains, "DeepContains", []string{"container", "elem"})
-	testCheck(c, DeepContains, false, "int is not a supported container", 5, nil)
-	testCheck(c, DeepContains, false, "bool is not a supported container", false, nil)
-	testCheck(c, DeepContains, false, "element is a int but expected a string", "container", 1)
-}
-
-func (s *CheckersS) TestDeepContainsVerifiesTypes(c *check.C) {
-	testInfo(c, DeepContains, "DeepContains", []string{"container", "elem"})
-	testCheck(c, DeepContains,
-		false, "container has items of type int but expected element is a string",
-		[...]int{1, 2, 3}, "foo")
-	testCheck(c, DeepContains,
-		false, "container has items of type int but expected element is a string",
-		[]int{1, 2, 3}, "foo")
-	// This looks tricky, DeepContains looks at _values_, not at keys
-	testCheck(c, DeepContains,
-		false, "container has items of type int but expected element is a string",
-		map[string]int{"foo": 1, "bar": 2}, "foo")
-}
-
-func (s *CheckersS) TestDeepContainsString(c *check.C) {
-	c.Assert("foo", DeepContains, "f")
-	c.Assert("foo", DeepContains, "fo")
-	c.Assert("foo", check.Not(DeepContains), "foobar")
-}
-
-func (s *CheckersS) TestDeepContainsCustomString(c *check.C) {
-	c.Assert(myString("foo"), DeepContains, myString("f"))
-	c.Assert(myString("foo"), DeepContains, myString("fo"))
-	c.Assert(myString("foo"), check.Not(DeepContains), myString("foobar"))
-	c.Assert("foo", DeepContains, myString("f"))
-	c.Assert("foo", DeepContains, myString("fo"))
-	c.Assert("foo", check.Not(DeepContains), myString("foobar"))
-	c.Assert(myString("foo"), DeepContains, "f")
-	c.Assert(myString("foo"), DeepContains, "fo")
-	c.Assert(myString("foo"), check.Not(DeepContains), "foobar")
-}
-
-func (s *CheckersS) TestDeepContainsArray(c *check.C) {
-	c.Assert([...]int{1, 2, 3}, DeepContains, 1)
-	c.Assert([...]int{1, 2, 3}, DeepContains, 2)
-	c.Assert([...]int{1, 2, 3}, DeepContains, 3)
-	c.Assert([...]int{1, 2, 3}, check.Not(DeepContains), 4)
-}
-
-func (s *CheckersS) TestDeepContainsSlice(c *check.C) {
-	c.Assert([]int{1, 2, 3}, DeepContains, 1)
-	c.Assert([]int{1, 2, 3}, DeepContains, 2)
-	c.Assert([]int{1, 2, 3}, DeepContains, 3)
-	c.Assert([]int{1, 2, 3}, check.Not(DeepContains), 4)
-}
-
-func (s *CheckersS) TestDeepContainsMap(c *check.C) {
-	c.Assert(map[string]int{"foo": 1, "bar": 2}, DeepContains, 1)
-	c.Assert(map[string]int{"foo": 1, "bar": 2}, DeepContains, 2)
-	c.Assert(map[string]int{"foo": 1, "bar": 2}, check.Not(DeepContains), 3)
-}
-
-func (s *CheckersS) TestDeepContainsUncomparableType(c *check.C) {
-	elem := myStruct{map[string]string{"k": "v"}}
-	containerArray := [...]myStruct{elem}
-	containerSlice := []myStruct{elem}
-	containerMap := map[string]myStruct{"foo": elem}
-	testInfo(c, DeepContains, "DeepContains", []string{"container", "elem"})
-	testCheck(c, DeepContains, true, "", containerArray, elem)
-	testCheck(c, DeepContains, true, "", containerSlice, elem)
-	testCheck(c, DeepContains, true, "", containerMap, elem)
 }
 
 type myStringer struct{ str string }
